requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
gunicorn==21.2.0
python-multipart==0.0.6
python-dotenv==1.0.0
numpy==1.24.3
opencv-python-headless==4.8.1.78
tensorflow==2.15.0
pydantic==2.5.2
pillow==10.0.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1